feat(be): add world join auth, move online world data to socket object
session token and world code is now checked when joining to make sure the user is allowed to join. if an existing connection from the same user ID exists, it is disconnected upon joining. need to add created player data to Redis database, as this doesn't work yet for some reason. packet middleware exists but does nothing - need to add checks for session expiry. need to test banlist and whitelist. need to get wifi to figure out why empty objects do not pass through the mongoose model create function. and need to make UI to show player list on frontend