feat(be): optimise join sequence, clean backend code
player count is now stored in a separate Redis key so that connected sockets do have to be counted. This reduces the operation for player count from O(N) to O(1). Also made same complexity reduction for kicking duplicate sockets when joining. Set the square for the player's spawn location upon joining to a player object. Need to add type distinguishing between different types of pieces.