feat(fe/be): send packet upon player move, load new chunks if needed
when you move the player, a player move packet is now sent to the server. the server will update the location of the player in the Redis database, and send any chunk load packets in case the player has entered a new chunk and is now in render distance of new chunks. a piece move packet is relayed to chunk subscribers, although the client does nothing with it right now. when the player on the client is moved, the viewport is clamped around their new position. Still need to move the piece in the local chunks, and then unload chunks that are no longer in view from the client, as updates will not be sent about this anymore. also need to edit chunk subscribers on the server-side player move handler, so that the player is no longer a subscriber of unloaded chunks.