fix: correct chunk subscriptions on player move
also removed duplicate local player entities on the frontend. the server was subscribing the player to the wrong chunks when diffing chunks upon player move. this caused the player to receive updates to chunks outside of their view, in turn creating ghost entities on the frontend, as it would not spawn entities entering the actual view under the assumption that they were already spawned. ugh